requirements.txt
tensorflow==1.9.0
opencv_python==4.1.0.25
xlwt==1.3.0
matplotlib==3.1.1
pandas==0.25.1
scipy==1.3.1
scikit_image==0.16.2
dicom2nifti==2.2.10
tensorlayer==1.11.0
nibabel==2.5.0
pycore==17.4.17
torchvision==0.4.0
~umpy==1.16.5
numpy==1.20.3
Pillow==8.2.0
scikit_learn==0.24.2
SimpleITK==2.0.2
skimage==0.0
